+++ NEWS.md
@@ -3,6 +3,7 @@
 * separationplot.spdur -> sepplot. `separationplot` is a standalone function so change the `spdur` version to a simple wrapper. 
 * Clean up NAMESPACE. Moved `stats` to imported package in description and removed all associated explicit namespace generic imports. 
 * Add package vignette.
+* Added file `init.c` with calls to `R_registerRoutines()` and `R_useDynamicSymbols()`; also use `.registration=TRUE` in `useDynLib` in `NAMESPACE`. R-devel (and R 3.4.0 in the future) `R CMD check` issues a NOTE for registration of routines, this is to avoid that note.  
 
 # spduration 0.15.1 (2016-05-12)